Initial file drop
All the HTML files I’ve created for the intro CS course over the past 8
years. This includes all of my old projects and labs. I reorganized the
subdirectories and moved the images to a central location, so there
might be some dead links. I moved all the txt files to data instead of
code, so some of those might be broken as well. You will also find my
rough lecture notes. The book is stored separately over in the
ThinkPython fork.